feat: Add toggle functionality for completing todos
This commit adds the ability to toggle the completion state of todos in the TodoCard component. It introduces a new action, toggleComplete, in the todoSlice reducer. The toggleComplete action takes the id of the todo as a payload and toggles the isCompleted property of the corresponding todo in the todos array in the state. The TodoCard component now uses the toggleComplete action to update the completion state when the checkbox is clicked.