feat: Add @radix-ui/react-select dependency and implement functionality to add new todos
This commit adds the @radix-ui/react-select dependency to the project by updating the package.json file. It also implements the functionality to add new todos in the TodoContainer component. The addTodo mutation is introduced in the baseApi module to send a POST request to the API with the new todo data. The TodoContainer component now uses the useAddTodoMutation hook to add new todos and updates the todos array in the state accordingly.